Changed the name of the package to Caching.jl
