+++ test/runtests.jl
@@ -1,6 +1,6 @@
 using Test
 using Random
-using LRUCaching
+using Caching
 
 # Function arguments
 _a_float = rand()
@@ -15,15 +15,15 @@ _an_int = rand(Int)
     bar(x; y=1) = x + y
 
     # Test caching of simple functions
-    foo_c1 = LRUCaching.MemoryCache(foo)
-    @test typeof(foo_c1) <: LRUCaching.AbstractCache
-    @test typeof(foo_c1) <: LRUCaching.MemoryCache
+    foo_c1 = MemoryCache(foo)
+    @test typeof(foo_c1) <: AbstractCache
+    @test typeof(foo_c1) <: MemoryCache
     @test foo(_a_float) == foo_c1(_a_float)
 
     # Test functions with keyword arguments
-    bar_c1 = LRUCaching.MemoryCache(bar)
-    @test typeof(bar_c1) <: LRUCaching.AbstractCache
-    @test typeof(bar_c1) <: LRUCaching.MemoryCache
+    bar_c1 = MemoryCache(bar)
+    @test typeof(bar_c1) <: AbstractCache
+    @test typeof(bar_c1) <: MemoryCache
     @test bar(_a_float) == bar_c1(_a_float)
     @test bar(_a_float; y=_a_float_2) == bar_c1(_a_float; y=_a_float_2)
 
@@ -31,8 +31,8 @@ _an_int = rand(Int)
     foo_m1 = @memcache foo			# no type annotations
 	foo_m2 = @memcache foo::Int		# standard type annotation
 	for _foo in (foo_m1, foo_m2)
-    	@test typeof(_foo) <: LRUCaching.AbstractCache
-    	@test typeof(_foo) <: LRUCaching.MemoryCache
+    	@test typeof(_foo) <: AbstractCache
+    	@test typeof(_foo) <: MemoryCache
 		@test _foo(_an_int) == foo(_an_int)	 # all should work with Int's
 		if !(_foo === foo_m1)
 		# `foo_m2` fails with arguments other than `Int`
@@ -57,18 +57,18 @@ end
     bar(x; y=1) = x + y
 
     # Test caching of simple functions
-    foo_c1 = LRUCaching.DiskCache(foo)
-    @test typeof(foo_c1) <: LRUCaching.AbstractCache
-    @test typeof(foo_c1) <: LRUCaching.DiskCache
-    @test typeof(foo_c1.memcache) <: LRUCaching.MemoryCache
+    foo_c1 = DiskCache(foo)
+    @test typeof(foo_c1) <: AbstractCache
+    @test typeof(foo_c1) <: DiskCache
+    @test typeof(foo_c1.memcache) <: MemoryCache
 	@test :filename in fieldnames(typeof(foo_c1))
     @test foo(_a_float) == foo_c1(_a_float)
 
     # Test functions with keyword arguments
-    bar_c1 = LRUCaching.DiskCache(bar)
-    @test typeof(bar_c1) <: LRUCaching.AbstractCache
-    @test typeof(bar_c1) <: LRUCaching.DiskCache
-    @test typeof(bar_c1.memcache) <: LRUCaching.MemoryCache
+    bar_c1 = DiskCache(bar)
+    @test typeof(bar_c1) <: AbstractCache
+    @test typeof(bar_c1) <: DiskCache
+    @test typeof(bar_c1.memcache) <: MemoryCache
 	@test :filename in fieldnames(typeof(bar_c1))
     @test bar(_a_float) == bar_c1(_a_float)
     @test bar(_a_float; y=_a_float_2) == bar_c1(_a_float; y=_a_float_2)
@@ -77,9 +77,9 @@ end
     foo_m1 = @diskcache foo			# no type annotations
 	foo_m2 = @diskcache foo::Int	# standard type annotation
 	for _foo in (foo_m1, foo_m2)
-    	@test typeof(_foo) <: LRUCaching.AbstractCache
-    	@test typeof(_foo) <: LRUCaching.DiskCache
-    	@test typeof(_foo.memcache) <: LRUCaching.MemoryCache
+    	@test typeof(_foo) <: AbstractCache
+    	@test typeof(_foo) <: DiskCache
+    	@test typeof(_foo.memcache) <: MemoryCache
 		@test :filename in fieldnames(typeof(_foo))
 		@test _foo(_an_int) == foo(_an_int)	 # all should work with Int's
 		if !(_foo === foo_m1)
